feat(stats): The Migration of The Backend charts: Total Blocks and Dynamic query dispatch (#1137)
#1080

It is a sequel to #1120 and the third installment in the Migrate Backend Charts franchise.

In this standalone sequel, we introduce reimagination of TotalBlocks counter and dynamic dispatch for chart query behaviour. TotalBlocks counter uses a custom logic with fallback to postgres catalog for preliminary block number estimation (to display when SQL query is not yet calculated), so we can't keep retrieving the data directly from DB. We need to allow dynamic query logic that is configurable for each chart.

While adding usage of dynamic logic for chart query, regular date range was replaced by universal range, because chart requests can include or omit any of the bounds.